% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps.R
\name{anjouan}
\alias{anjouan}
\title{Anjouan}
\usage{
anjouan(pref = FALSE, city = TRUE)
}
\arguments{
\item{pref}{Choose to map with prefecture area ("pref" = TRUE) or without prefecture area ("pref" = FALSE)}

\item{city}{Include all Anjouan cities as point features (\code{TRUE}) or exclude them (\code{FALSE}). Default is \code{TRUE}.}
}
\value{
The data set used is in \code{sf} format
}
\description{
Draw a map for Anjouan Island
}
\examples{
## Map Anjouan Island.
anjouan ()
## Map Anjouan Island with prefecture area.
anjouan (pref = TRUE)

}
