% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps.R
\name{comoros}
\alias{comoros}
\title{Comoro Islands}
\usage{
comoros(x = "country", pref = FALSE, city = FALSE)
}
\arguments{
\item{x}{Name of the data set to use. The default is \code{comoros()}, It draws Comoro Islands as one object without commune.}

\item{pref}{Choose to map with prefecture area ("pref" = TRUE) or without prefecture area ("pref" = FALSE)}

\item{city}{Include all cities as point features (\code{TRUE}) or exclude them (\code{FALSE}). Default is \code{FALSE}.}
}
\value{
The data set used is in \code{sf} format
}
\description{
Draw  maps of Comoros Islands.
}
\details{
\code{comoros()}uses the sf package to plot by plotting the geometry rather than every column and leave the plot pref ready for overplotting with other data
}
\examples{
## Map Comoro Islands as one object without prefecture area
comoros()
## Map Comoro Islands as one object with prefecture area
comoros(x="country",pref=TRUE)
## Map Comoros Islands as 3 object (Grande Comore, Anjouan, Mohéli) without prefecture area.
comoros(x="island",pref=FALSE)
## Map Comoros Islands as 3 object (Grande Comore, Anjouan, Mohéli) with prefecture area
comoros(x="island",pref=TRUE)

}
