% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps.R
\name{grandeComore}
\alias{grandeComore}
\title{Grande Comore}
\usage{
grandeComore(pref = FALSE, city = TRUE)
}
\arguments{
\item{pref}{Choose to map with prefecture area ("pref" = TRUE) or without prefecture area ("pref" = FALSE)}

\item{city}{Include all Grande Comore cities as point features (\code{TRUE}) or exclude them (\code{FALSE}). Default is \code{TRUE}.}
}
\value{
The data set used is in \code{sf} format
}
\description{
Draw a map for Grande Comore Island
}
\examples{
## Map Grande Comore Island
grandeComore ()
## Map Grande Comore with prefecture area
grandeComore (pref = TRUE)

}
