% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps.R
\name{moheli}
\alias{moheli}
\title{Moheli}
\usage{
moheli(pref = FALSE, city = TRUE)
}
\arguments{
\item{pref}{Choose to map with prefecture area ("pref" = TRUE) or without prefecture area ("pref" = FALSE)}

\item{city}{Include all Mohéli cities as point features (\code{TRUE}) or exclude them (\code{FALSE}). Default is \code{TRUE}.}
}
\value{
The data set used is in \code{sf} format
}
\description{
Draw a map for Moheli Islands
}
\examples{
## Map Moheli Island
moheli ()
## Map Moheli Island with prefecture area
moheli (pref = TRUE)
}
